requirements.txt
faker
pillow
easyocr
pandas
